requirements-dev.txt
# Additional libraries required to build Autolysis

# Development
wheel
flake8 >= 2.5           # To detect [flake8] section in setup.cfg
coverage
Sphinx
cryptography
PyYAML

# Documentation
sphinx_rtd_theme

# Testing
nose >= 1.3
markdown
PyMySQL
psycopg2
